// vuejs repo/src/api/generated/models/AuthenticationTokenObjectResponseApiResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { AuthenticationTokenObjectResponse } from './AuthenticationTokenObjectResponse'
export type AuthenticationTokenObjectResponseApiResponse = {
  requestId?: string
  isSuccess?: boolean
  message?: string
  data?: AuthenticationTokenObjectResponse
}